refactor(GroupConcern): Move useful methods to concern and rename it
