+++ api/app/controllers/groups_controller.rb
@@ -1,5 +1,5 @@
 class GroupsController < ApplicationController
-  include GroupAdminConcern
+  include GroupAuthorizable
 
   before_action :set_group, only: %i[show update destroy members]
   before_action :authorize_group_admin!, only: %i[update destroy]
