+++ api/app/controllers/sessions_controller.rb
@@ -1,4 +1,6 @@
 class SessionsController < ApplicationController
+  include GroupAuthorizable
+
   before_action :set_session, only: %i[show update destroy]
   before_action :authorize_group_member!, only: %i[show create]
   before_action :authorize_creator_or_group_admin!, only: %i[update destroy]
@@ -64,25 +66,15 @@ class SessionsController < ApplicationController
   end
 
   def session_params
-    params.require(:session).permit(:name, :description, :location, :meeting_link, :start_time, :end_time, :group_id)
-  end
-
-  def authorize_group_member!
-    group = @session ? @session.group : Group.find(session_params[:group_id])
-
-    member = Member.find_by(user: current_user, group:)
-
-    return if member
-
-    render json: { message: 'No estás autorizado para realizar esta acción' }, status: :unauthorized
-  end
-
-  def authorize_creator_or_group_admin!
-    is_admin = @session.group.admin?(current_user)
-    is_creator = @session.creator == current_user
-
-    return if is_admin || is_creator
-
-    render json: { message: 'No estás autorizado para realizar esta acción' }, status: :unauthorized
+    params.require(:session)
+          .permit(
+            :name,
+            :description,
+            :location,
+            :meeting_link,
+            :start_time,
+            :end_time,
+            :group_id
+          )
   end
 end
